Tests/LibWeb/Text/expected/non-html-mixed-case-element-name-selector-matching.txt

✅ Pass: Selector match for SVG element textPath.
✅ Pass: Selector match for SVG element clipPath.
✅ Pass: Selector match for SVG element foreignObject.
✅ Pass: Selector match for SVG element radialGradient.
